RAP-2123 Enable mixing in browser version of Disposable
